requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
channels-redis==4.1.0
constantly==15.1.0
crispy-bootstrap4==2022.1
cryptography==40.0.2
daphne==3.0.2
Django==4.2.1
django-crispy-forms==2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
redis==4.5.5
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.1
zope.interface==6.0
psycopg2==2.9.6
python-decouple==3.8
django-elasticsearch-dsl==7.3